# PYTHON-4292 Improve TLS read performance