PYTHON-4292 Improve TLS read performance

Use a socket timeout based approach for cancelling pending operations. This results in >=2x better throughput when reading large TLS responses from multiple threads.

Before:
```
time python bench-100-threads-find-large.py
Python: 3.13.0, PyMongo: 4.11.0.dev0, MongoDB: 8.0.1, auth: True, tls: True
python bench-100-threads-find-large.py  15.05s user 53.82s system 256% cpu 26.805 total
```
After:
```
time python bench-100-threads-find-large.py
Python: 3.13.0, PyMongo: 4.11.0.dev0, MongoDB: 8.0.1, auth: True, tls: True
python bench-100-threads-find-large.py  9.31s user 23.28s system 244% cpu 13.327 total
```