@ShaneHarvey 2024-12-02
Update I also confirmed the perf improvement on rhel8.7-small host (updated the benchmark to use 20 threads to time's sake). Before:
```
$ time python bench-20-threads-find-large.py
Python: 3.13.0, PyMongo: 4.11.0.dev0, MongoDB: 8.0.4-rc0, auth: True, tls: True

real	0m13.271s
user	0m12.706s
sys	0m10.766s
```
After:
```
$ time python bench-20-threads-find-large.py
Python: 3.13.0, PyMongo: 4.11.0.dev0, MongoDB: 8.0.4-rc0, auth: True, tls: True

real	0m7.784s
user	0m7.708s
sys	0m4.563s
```